Source Code/FusionNet-GLD/backend/requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
joblib
pandas